requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.56
botocore==1.35.56
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
jwt==1.3.1
mysql-connector-python==9.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
requests~=2.32.3